* fixed the type representation of the durations in the appenders (32bit integer is not big enough for milliseconds since 1970) * fixed a missing renaming of setLimitReliability in the merge playground app

+++ libLogging/src/logging/ConsoleAppender.cpp
@@ -11,6 +11,7 @@
 #include <sstream>
 #include <iomanip>
 #include <chrono>
+#include <cstdint>
 
 using std::cout;
 using std::ostringstream;
@@ -33,9 +34,9 @@ void ConsoleAppender::log(const loglevel logLevel, const string loggerName, cons
 string ConsoleAppender::getCurrentTime()
 {
 	system_clock::time_point now = system_clock::now();
-	duration<int, std::ratio<1>> seconds = duration_cast<duration<int, std::ratio<1>>>(now.time_since_epoch());
-	duration<int, std::milli> milliseconds = duration_cast<duration<int, std::milli>>(now.time_since_epoch());
-	duration<int, std::milli> msec = milliseconds - seconds;
+	duration<int64_t, std::ratio<1>> seconds = duration_cast<duration<int64_t, std::ratio<1>>>(now.time_since_epoch());
+	duration<int64_t, std::milli> milliseconds = duration_cast<duration<int64_t, std::milli>>(now.time_since_epoch());
+	duration<int64_t, std::milli> msec = milliseconds - seconds;
 
 	std::time_t t_now = system_clock::to_time_t(now);
 	struct tm* tm_now = std::localtime(&t_now);
